+++ const_format/tests/misc_tests_modules.rs
@@ -8,9 +8,12 @@ pub const NOT_CF: usize = 13;
 pub const _ASSERT_NOT_CF: [(); 13] = [(); const_format::NOT_CF];
 
 mod misc_tests {
-    #[cfg(feature = "assert")]
+    #[cfg(feature = "assertc")]
     #[cfg(not(feature = "only_new_tests"))]
-    mod asserts;
+    mod assertc_tests;
+
+    #[cfg(feature = "assertcp")]
+    mod assertcp_tests;
 
     #[cfg(feature = "fmt")]
     #[cfg(not(feature = "only_new_tests"))]
@@ -24,7 +27,7 @@ mod misc_tests {
     #[cfg(not(feature = "only_new_tests"))]
     mod derive_tests;
 
-    #[cfg(feature = "assert")]
+    #[cfg(feature = "assertc")]
     #[cfg(not(feature = "only_new_tests"))]
     mod equality_tests;
 
@@ -35,7 +38,7 @@ mod misc_tests {
     #[cfg(not(feature = "only_new_tests"))]
     mod impl_fmt_macro_tests;
 
-    #[cfg(feature = "assert")]
+    #[cfg(feature = "assertc")]
     #[cfg(not(feature = "only_new_tests"))]
     mod shared_cp_macro_tests;